Refactor headers in panda/board/*.h
