+++ board/fake_stm.c
@@ -1,9 +1,5 @@
 // minimal code to fake a panda for tests
-#include <stdio.h>
-#include <stdint.h>
-#include <stdlib.h>
-
-#include "utils.h"
+#include "fake_stm.h"
 
 #define CANFD
 #define ALLOW_DEBUG
@@ -12,9 +8,13 @@
 #define ENTER_CRITICAL() 0
 #define EXIT_CRITICAL() 0
 
-void print(const char *a);
+void print(const char *a) {
+  printf("%s", a);
+}
 
-void puth(unsigned int i);
+void puth(unsigned int i) {
+  printf("%u", i);
+}
 
 typedef struct {
   uint32_t CNT;
@@ -24,4 +24,6 @@ TIM_TypeDef timer;
 TIM_TypeDef *MICROSECOND_TIMER = &timer;
 uint32_t microsecond_timer_get(void);
 
-uint32_t microsecond_timer_get(void);
+uint32_t microsecond_timer_get(void) {
+  return MICROSECOND_TIMER->CNT;
+}
