+++ board/stm32f4/stm32f4_config.h
@@ -34,7 +34,7 @@
 #include "comms_definitions.h"
 
 #ifndef BOOTSTUB
-  #include "main_definitions.h"
+  #include "main_declarations.h"
 #else
   #include "bootstub_declarations.h"
 #endif
